Revert "Updating Health to be Float instead of Int"
This reverts commit f7329e342263bf8a83f3cc411c728857741edd58.

+++ TopDownTPGame/Assets/Scripts/Ability Scripts/Spawns/Type_2_IceWall.cs
@@ -88,9 +88,9 @@ namespace Ability_Scripts.Spawns
 
         #region Utils
 
-        private void HandleHealthAndDamage(float startHealth, float currentHealth, float maxHealth)
+        private void HandleHealthAndDamage(int startHealth, int currentHealth, int maxHealth)
         {
-            _material.color = Color.Lerp(_minHealthColor, _fullHealthColor, currentHealth / maxHealth);
+            _material.color = Color.Lerp(_minHealthColor, _fullHealthColor, (float)currentHealth / maxHealth);
             if (currentHealth <= 0)
             {
                 DestroyWall();
